requirements/dev.txt
-r production.txt

django-debug-toolbar
Werkzeug
django-extensions
fabric
